LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/html/canvas/element/manual/wide-gamut-canvas/canvas-createImageBitmap-e_srgb-expected.txt


FAIL createImageBitmap in e-sRGB from an opaque sRGB HTMLImageElement (avif) with resize. promise_test: Unhandled rejection with value: object "EncodingError: Loading error."
FAIL createImageBitmap in e-sRGB from an opaque sRGB HTMLImageElement (bmp) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB HTMLImageElement (gif) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB HTMLImageElement (ico) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB HTMLImageElement (jpg) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB HTMLImageElement (png) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB HTMLImageElement (webp) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from a transparent sRGB HTMLImageElement (avif) with resize. promise_test: Unhandled rejection with value: object "EncodingError: Loading error."
FAIL createImageBitmap in e-sRGB from a transparent sRGB HTMLImageElement (bmp) with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.88623 +/- 0.025 but got 225
FAIL createImageBitmap in e-sRGB from a transparent sRGB HTMLImageElement (ico) with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.88623 +/- 0.025 but got 225
FAIL createImageBitmap in e-sRGB from a transparent sRGB HTMLImageElement (png) with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.88623 +/- 0.025 but got 225
FAIL createImageBitmap in e-sRGB from a transparent sRGB HTMLImageElement (webp) with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.88623 +/- 0.025 but got 227
FAIL createImageBitmap in e-sRGB from a sRGB SVG image with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
NOTRUN createImageBitmap in e-sRGB from a sRGB HTMLVideoElement with resize. video/ogg not supported
FAIL createImageBitmap in e-sRGB from an opaque sRGB HTMLCanvasElement with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque e-sRGB HTMLCanvasElement with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from a transparent sRGB HTMLCanvasElement with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from a transparent e-sRGB HTMLCanvasElement with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB Blob (avif) with resize. promise_test: Unhandled rejection with value: object "InvalidStateError: Cannot decode the data in the argument to createImageBitmap"
FAIL createImageBitmap in e-sRGB from an opaque sRGB Blob (bmp) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB Blob (gif) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB Blob (ico) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB Blob (jpg) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB Blob (png) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB Blob (webp) with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from a transparent sRGB Blob (avif) with resize. promise_test: Unhandled rejection with value: object "InvalidStateError: Cannot decode the data in the argument to createImageBitmap"
FAIL createImageBitmap in e-sRGB from a transparent sRGB Blob (bmp) with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.88623 +/- 0.025 but got 225
FAIL createImageBitmap in e-sRGB from a transparent sRGB Blob (ico) with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.88623 +/- 0.025 but got 225
FAIL createImageBitmap in e-sRGB from a transparent sRGB Blob (png) with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.88623 +/- 0.025 but got 225
FAIL createImageBitmap in e-sRGB from a transparent sRGB Blob (webp) with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.88623 +/- 0.025 but got 227
FAIL createImageBitmap in e-sRGB from color managed Blob with resize. blobPixelFormat: uint8, transparency: false promise_test: Unhandled rejection with value: object "TypeError: testCanvas.convertToBlob is not a function. (In 'testCanvas.convertToBlob({
            quality: 1,
            type: 'image/png',
            pixelFormat,
        })', 'testCanvas.convertToBlob' is undefined)"
FAIL createImageBitmap in e-sRGB from color managed Blob with resize. blobPixelFormat: uint8, transparency: true promise_test: Unhandled rejection with value: object "TypeError: testCanvas.convertToBlob is not a function. (In 'testCanvas.convertToBlob({
            quality: 1,
            type: 'image/png',
            pixelFormat,
        })', 'testCanvas.convertToBlob' is undefined)"
FAIL createImageBitmap in e-sRGB from color managed Blob with resize. blobPixelFormat: uint16, transparency: false promise_test: Unhandled rejection with value: object "TypeError: testCanvas.convertToBlob is not a function. (In 'testCanvas.convertToBlob({
            quality: 1,
            type: 'image/png',
            pixelFormat,
        })', 'testCanvas.convertToBlob' is undefined)"
FAIL createImageBitmap in e-sRGB from color managed Blob with resize. blobPixelFormat: uint16, transparency: true promise_test: Unhandled rejection with value: object "TypeError: testCanvas.convertToBlob is not a function. (In 'testCanvas.convertToBlob({
            quality: 1,
            type: 'image/png',
            pixelFormat,
        })', 'testCanvas.convertToBlob' is undefined)"
FAIL createImageBitmap in e-sRGB from an opaque sRGB ImageData with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque e-sRGB ImageData with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from a transparent sRGB ImageData with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from a transparent e-sRGB ImageData with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB ImageBitmap with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque e-sRGB ImageBitmap with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from a transparent sRGB ImageBitmap with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from a transparent e-sRGB ImageBitmap with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque sRGB OffscreenCanvas with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from an opaque e-sRGB OffscreenCanvas with resize. assert_approx_equals: This pixel should be e-sRGB like red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from a transparent sRGB OffscreenCanvas with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.607422 +/- 0.025 but got 155
FAIL createImageBitmap in e-sRGB from a transparent e-sRGB OffscreenCanvas with resize. assert_approx_equals: This pixel should be e-sRGB transparent red. expected 0.607422 +/- 0.025 but got 155
